% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoZUR_count.R
\docType{data}
\name{EcoZUR_count}
\alias{EcoZUR_count}
\title{Count data from EcoZUR study, example data used in examples and vignettes}
\format{
A data frame with with 358 rows and 229 columns. Each row is an observation and each column is a microbial taxon.
}
\source{
Jesser et al. (2023). "Why are so many enteric pathogen infections asymptomatic? Pathogen and gut microbiome characteristics associated with diarrhea symptoms and carriage of diarrheagenic E. coli in northern Ecuador." Gut Microbes 15(2).
}
\usage{
EcoZUR_count
}
\description{
A dataset containing the count matrix from the EcoZur study.
}
\keyword{datasets}
