% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoZUR_meta.R
\docType{data}
\name{EcoZUR_meta}
\alias{EcoZUR_meta}
\title{Metadata from EcoZUR study, example data used in examples and vignettes}
\format{
A data frame with 358 rows and 5 variables:
\describe{
  \item{Diarrhea}{Indicator of whether sample is from a participant with diarrhea (Case) or a participant without diarrhea (control) (character)}
  \item{sex}{Sex of participant (character)}
  \item{age_months}{Age in months of participant (integer)}
  \item{urban}{How urban of a population the sample is from, from 0 to 5 (integer)}
  \item{site}{Study site (character)}
}
}
\source{
Jesser et al. (2023). "Why are so many enteric pathogen infections asymptomatic? Pathogen and gut microbiome characteristics associated with diarrhea symptoms and carriage of diarrheagenic E. coli in northern Ecuador." Gut Microbes 15(2).
}
\usage{
EcoZUR_meta
}
\description{
A dataset containing metadata from the EcoZur study.
}
\keyword{datasets}
