% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndFit.R
\name{ndFit}
\alias{ndFit}
\title{Fit niceday model}
\usage{
ndFit(
  W,
  A,
  X,
  alpha = 0.05,
  uniform_CI = TRUE,
  d = 0.1,
  gtrunc = 0.01,
  bs_rep = 1e+05,
  num_crossval_folds = 10,
  num_crossfit_folds = 10,
  enforce_pos_reg = FALSE,
  sl.lib.pi = c("SL.mean", "SL.lm", "SL.glm.binom", "SL.xgboost.binom"),
  sl.lib.m = c("SL.mean", "SL.lm", "SL.glm.qpois", "SL.xgboost.pois"),
  sl.lib.q = sl.lib.pi,
  adjust_covariates = TRUE,
  nuis = NULL,
  verbose = FALSE,
  cross_fit = TRUE
)
}
\arguments{
\item{W}{Matrix of multivariate outcome variables.}

\item{A}{Binary covariate of interest.}

\item{X}{Matrix of covariates to adjust for.}

\item{alpha}{Desired asymptotic type I error rate. Default is `0.05`.}

\item{uniform_CI}{?? Default is `TRUE`.}

\item{d}{Parameter for smoothed median centering. Default is `0.1`.}

\item{gtrunc}{?? Default is `0.01`.}

\item{bs_rep}{Number of replicates for bootstrap sampling of uniform confidence intervals. Default is `1e5`.}

\item{num_crossval_folds}{Number of folds for cross-validation. Default is `10`.}

\item{num_crossfit_folds}{Number of folds for cross-fitting. Default is `10`.}

\item{enforce_pos_reg}{Should estimates of \eqn{E[W_j|A=a,X]} to forced to be strictly positive? Default is `FALSE`.}

\item{sl.lib.pi}{Libraries used to estimate ???. Default is `c("SL.mean", "SL.lm", "SL.glm.binom", "SL.xgboost.binom")`.}

\item{sl.lib.m}{Libraries used to estimate ???. Default is `c("SL.mean", "SL.lm", "SL.glm.qpois", "SL.xgboost.pois")`.}

\item{sl.lib.q}{Libraries used to estimate ???. Default is the input to `sl.lib.pi`.}

\item{adjust_covariates}{Should covariates be adjusted for in the estimator? Default is `TRUE`.}

\item{nuis}{??. Default is `NULL`.}

\item{verbose}{Do you want to receive updates as this function runs? Default is `FALSE`.}

\item{cross_fit}{Should cross-fitting be run? Default is `TRUE`.}
}
\value{
A list containing elements `noadjust`, `adjust`, `nuis`, `cf_nuis`, and `variance`. `noadjust` gives unadjusted parameter estimates.
`adjust` gives covariate adjusted parameter estimates. `nuis` gives estimates of nuisance parameters. Add a description of the rest here!
}
\description{
Fit niceday model
}
\examples{
# add example here!

}
