% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceday-package.R
\docType{package}
\name{niceday}
\alias{niceday-package}
\alias{niceday}
\title{niceday: Nonparametrically Identified (de-Confounded) Estimands for Difference Abundance, Yippee!}
\description{
A nonparametric method to estimate fold-differences associated with binary covariates for multivariate outcomes.
}
\author{
\strong{Maintainer}: Amy D Willis \email{adwillis@uw.edu} (\href{https://orcid.org/0000-0002-2802-4317}{ORCID})

Authors:
\itemize{
  \item Grant Hopkins
  \item Sarah Teichman
}

}
\keyword{internal}
