% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ndFit.R
\name{print.ndFit}
\alias{print.ndFit}
\title{Print function}
\usage{
\method{print}{ndFit}(x, n = 20, ...)
}
\arguments{
\item{x}{Object of class \code{ndFit}}

\item{n}{The number of coefficient estimates to be printed (ordered by largest absolute value to smallest)}

\item{...}{No optional arguments are accepted at this time.}
}
\value{
\code{NULL}. Displays printed model summary.
}
\description{
Print function
}
