% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndFit.R
\name{ndFit}
\alias{ndFit}
\title{Fit niceday model}
\usage{
ndFit(
  W,
  A,
  X = ~1,
  data,
  alpha = 0.05,
  uniform_CI = TRUE,
  d = 0.1,
  gtrunc = 0.01,
  bs_rep = 1e+05,
  num_crossval_folds = 10,
  num_crossfit_folds = 10,
  enforce_pos_reg = FALSE,
  sl.lib.pi = c("SL.mean", "SL.lm", "SL.glm.binom", "SL.xgboost.binom"),
  sl.lib.m = c("SL.mean", "SL.lm", "SL.glm.qpois", "SL.xgboost.pois"),
  sl.lib.q = sl.lib.pi,
  nuis = NULL,
  verbose = FALSE,
  cross_fit = TRUE
)
}
\arguments{
\item{W}{Matrix of multivariate outcome variables.}

\item{A}{Formula indicating binary covariate of interest.}

\item{X}{Formula indicating covariates to adjust for. Default is no adjustment, `~ 1`}

\item{data}{Data frame containing any covariates in formulas for `A` and `X`.}

\item{alpha}{Desired asymptotic type I error rate. Default is `0.05`.}

\item{uniform_CI}{Generate simultaneous confidence intervals across al categories? Default is `TRUE`.}

\item{d}{Parameter for smoothed median centering. Default is `0.1`.}

\item{gtrunc}{Truncation parameter passed to `est_nuis`, bounding the estimated propensity scores away from `0` and `1` by `gtrunc`. Default is `0.01`.}

\item{bs_rep}{Number of replicates for bootstrap sampling of uniform confidence intervals. Default is `1e5`.}

\item{num_crossval_folds}{Number of folds for cross-validation. Default is `10`.}

\item{num_crossfit_folds}{Number of folds for cross-fitting. Default is `10`.}

\item{enforce_pos_reg}{Should estimates of \eqn{E[W_j|A=a,X]} to forced to be strictly positive? Default is `FALSE`.}

\item{sl.lib.pi}{Libraries used to estimate the propensity score nuisance function. Default is `c("SL.mean", "SL.lm", "SL.glm.binom", "SL.xgboost.binom")`.}

\item{sl.lib.m}{Libraries used to estimate conditional mean mu_j nuisance functions. Default is `c("SL.mean", "SL.lm", "SL.glm.qpois", "SL.xgboost.pois")`.}

\item{sl.lib.q}{Libraries used to estimate conditional probabilities on nonzero observations. Default is the input to `sl.lib.pi`.}

\item{nuis}{Optional list of estimated nuisance functions (from function `est_nuis()`). Default is `NULL`.}

\item{verbose}{Do you want to receive updates as this function runs? Default is `FALSE`.}

\item{cross_fit}{Should cross-fitting be run? Default is `TRUE`.}
}
\value{
A list containing elements `coef`, `nuisances`, `crossfit_nuisances`, `variance`, and `call`. `coef` contains estimates,
standard errors, and confidence intervals for all categories. `nuisances` and `crossfit_nuisances` contain estimated nuisance
parameters and nuisance parameters from the crossfitting procedure (included when adjustment covariates are included).
`variance` contains the estimated covariance matrix. `call` contains the call to `ndFit()`.
}
\description{
Fit niceday model
}
\examples{
data(EcoZUR_meta)
data(EcoZUR_count)
ndFit(W = EcoZUR_count[, 1:50], # consider only the first 50 taxa to run quickly
      data = EcoZUR_meta,
      A = ~ Diarrhea,
      X = ~ sex + age_months,
      num_crossval_folds = 2, # use more folds in practice
      num_crossfit_folds = 2, # for cross validation and cross fitting
      sl.lib.pi = c("SL.mean"), # choosing single learner for the example to run quickly,
      sl.lib.m = c("SL.mean"))  # in practice would use other options as well

}
